requirements.txt
instructor
pydantic
pydantic-settings
python-dotenv
fastapi[all]
octoai-sdk
gunicorn
anthropic
openai
fal-client